refactor(uniswap): refactor add add utility functions
